Fixed BLL with Validation

+++ Progame.HospitalAPI.BLL/PatientService.cs
@@ -1,4 +1,5 @@
 ﻿using ProGame.HospitalAPI.BLL.Interfaces;
+using ProGame.HospitalAPI.BLL.Validation;
 using ProGame.HospitalAPI.Common.Entities;
 using ProGame.HospitalAPI.DAL.Interfaces;
 using System;
@@ -18,14 +19,57 @@ namespace Progame.HospitalAPI.BLL
             _patientDAO = patientDAO;
         }
 
-        public int Add(Patient patient)
+        public ActionResult<int?> Add(Patient patient)
         {
-            return _patientDAO.Add(patient);
+            var validator = new PatientValidator();
+            var validationResult = validator.Validate(patient);
+
+            if (validationResult.IsValid)
+            {
+                int? id = null;
+                try
+                {
+                    id = _patientDAO.Add(patient);
+                }
+                catch (Exception e)
+                {
+                    return new ActionResult<int?>(null, new List<string>()
+                    {
+                        e.Message
+                    });
+                }
+                return new ActionResult<int?>(id, new List<string>());
+            }
+            else
+            {
+                return new ActionResult<int?>(null, validationResult.Errors.Select(e => e.ErrorMessage).ToList());
+            }
         }
 
-        public void Delete(Patient patient)
+        public ActionResult<bool> Delete(Patient patient)
         {
-            _patientDAO.Delete(patient);
+            var validator = new PatientValidator();
+            var validationResult = validator.Validate(patient);
+
+            if (validationResult.IsValid)
+            {
+                try
+                {
+                    _patientDAO.Delete(patient);
+                }
+                catch (Exception e)
+                {
+                    return new ActionResult<bool>(false, new List<string>()
+                    {
+                        e.Message
+                    });
+                }
+                return new ActionResult<bool>(true, new List<string>());
+            }
+            else
+            {
+                return new ActionResult<bool>(false, validationResult.Errors.Select(e => e.ErrorMessage).ToList());
+            }
         }
 
         public IEnumerable<Patient> GetAll()
@@ -38,9 +82,30 @@ namespace Progame.HospitalAPI.BLL
             return _patientDAO.GetById(id);
         }
 
-        public void Update(Patient patient)
+        public ActionResult<bool> Update(Patient patient)
         {
-            _patientDAO.Update(patient);
+            var validator = new PatientValidator();
+            var validationResult = validator.Validate(patient);
+
+            if (validationResult.IsValid)
+            {
+                try
+                {
+                    _patientDAO.Update(patient);
+                }
+                catch (Exception e)
+                {
+                    return new ActionResult<bool>(false, new List<string>()
+                    {
+                        e.Message
+                    });
+                }
+                return new ActionResult<bool>(true, new List<string>());
+            }
+            else
+            {
+                return new ActionResult<bool>(false, validationResult.Errors.Select(e => e.ErrorMessage).ToList());
+            }
         }
     }
 }
